Fix cvwait hangs

+++ src/kernel/emulation/linux/psynch/psynch_cvwait.c
@@ -6,7 +6,7 @@
 #include "../../../../lkm/api.h"
 #include "../simple.h"
 
-long sys_psynch_cvwait(void* cv, uint32_t cvgen, uint32_t cvugen, void* mutex, uint32_t mgen,
+long sys_psynch_cvwait(void* cv, uint32_t cvgen, uint32_t cvugen, void* mutex, uint64_t mgen,
 		uint32_t ugen, uint64_t sec, uint32_t usec)
 {
 	struct psynch_cvwait_args args = {
