+++ src/kernel/emulation/linux/psynch/psynch_cvwait.h
@@ -2,7 +2,7 @@
 #define LINUX_PSYNCH_CVWAIT_H
 #include <stdint.h>
 
-long sys_psynch_cvwait(void* cv, uint32_t cvgen, uint32_t cvugen, void* mutex, uint32_t mgen,
+long sys_psynch_cvwait(void* cv, uint32_t cvgen, uint32_t cvugen, void* mutex, uint64_t mgen,
 		uint32_t ugen, uint64_t sec, uint32_t usec);
 
 #endif
